+++ tests/integration/buildcmd/test_build_cmd.py
@@ -1,19 +1,16 @@
 import logging
 import os
 import random
-import shutil
 import sys
 from pathlib import Path
 from typing import Set
 from unittest import skipIf
 
-import jmespath
 import docker
+import jmespath
 import pytest
 from parameterized import parameterized, parameterized_class
 
-from samcli.commands.build.utils import MountMode
-from samcli.lib.utils import osutils
 from samcli.yamlhelper import yaml_parse
 from tests.testing_utils import (
     IS_WINDOWS,
@@ -25,219 +22,20 @@ from tests.testing_utils import (
     SKIP_DOCKER_TESTS,
     SKIP_DOCKER_BUILD,
     SKIP_DOCKER_MESSAGE,
-    run_command_with_input,
 )
 from .build_integ_base import (
     BuildIntegBase,
     DedupBuildIntegBase,
     CachedBuildIntegBase,
-    BuildIntegRubyBase,
     NestedBuildIntegBase,
     IntrinsicIntegBase,
-    BuildIntegNodeBase,
-    BuildIntegGoBase,
-    BuildIntegProvidedBase,
-    BuildIntegPythonBase,
-    BuildIntegJavaBase,
-    BuildIntegEsbuildBase,
 )
+from ..buildcmd_interpreted.test_build_python import BuildIntegPythonBase
 
 LOG = logging.getLogger(__name__)
 
 # SAR tests require credentials. This is to skip running the test where credentials are not available.
 SKIP_SAR_TESTS = RUNNING_ON_CI and RUNNING_TEST_FOR_MASTER_ON_CI and not RUN_BY_CANARY
-
-
-@skipIf(
-    # Hits public ECR pull limitation, move it to canary tests
-    (not RUN_BY_CANARY and not CI_OVERRIDE),
-    "Skip build tests on windows when running in CI unless overridden",
-)
-class TestBuildCommand_PythonFunctions_Images(BuildIntegBase):
-    template = "template_image.yaml"
-
-    EXPECTED_FILES_PROJECT_MANIFEST: Set[str] = set()
-
-    FUNCTION_LOGICAL_ID_IMAGE = "ImageFunction"
-
-    @parameterized.expand([("3.7", False), ("3.8", False), ("3.9", False)])
-    @pytest.mark.flaky(reruns=3)
-    def test_with_default_requirements(self, runtime, use_container):
-        _tag = f"{random.randint(1,100)}"
-        overrides = {
-            "Runtime": runtime,
-            "Handler": "main.handler",
-            "DockerFile": "Dockerfile",
-            "Tag": _tag,
-        }
-        cmdlist = self.get_command_list(use_container=use_container, parameter_overrides=overrides)
-
-        LOG.info("Running Command: ")
-        LOG.info(cmdlist)
-        command_result = run_command(cmdlist, cwd=self.working_dir)
-        self.assertEqual(command_result.process.returncode, 0)
-
-        self._verify_image_build_artifact(
-            self.built_template,
-            self.FUNCTION_LOGICAL_ID_IMAGE,
-            "ImageUri",
-            f"{self.FUNCTION_LOGICAL_ID_IMAGE.lower()}:{_tag}",
-        )
-
-        expected = {"pi": "3.14"}
-        self._verify_invoke_built_function(
-            self.built_template, self.FUNCTION_LOGICAL_ID_IMAGE, self._make_parameter_override_arg(overrides), expected
-        )
-
-    @parameterized.expand([("3.7", False), ("3.8", False), ("3.9", False)])
-    @pytest.mark.flaky(reruns=3)
-    def test_with_dockerfile_extension(self, runtime, use_container):
-        _tag = f"{random.randint(1,100)}"
-        overrides = {
-            "Runtime": runtime,
-            "Handler": "main.handler",
-            "DockerFile": "Dockerfile.production",
-            "Tag": _tag,
-        }
-        cmdlist = self.get_command_list(use_container=use_container, parameter_overrides=overrides)
-
-        LOG.info("Running Command: ")
-        LOG.info(cmdlist)
-        command_result = run_command(cmdlist, cwd=self.working_dir)
-        self.assertEqual(command_result.process.returncode, 0)
-
-        self._verify_image_build_artifact(
-            self.built_template,
-            self.FUNCTION_LOGICAL_ID_IMAGE,
-            "ImageUri",
-            f"{self.FUNCTION_LOGICAL_ID_IMAGE.lower()}:{_tag}",
-        )
-
-        expected = {"pi": "3.14"}
-        self._verify_invoke_built_function(
-            self.built_template, self.FUNCTION_LOGICAL_ID_IMAGE, self._make_parameter_override_arg(overrides), expected
-        )
-
-    @pytest.mark.flaky(reruns=3)
-    def test_intermediate_container_deleted(self):
-        _tag = f"{random.randint(1, 100)}"
-        overrides = {
-            "Runtime": "3.9",
-            "Handler": "main.handler",
-            "DockerFile": "Dockerfile",
-            "Tag": _tag,
-        }
-        cmdlist = self.get_command_list(use_container=False, parameter_overrides=overrides)
-
-        LOG.info("Running Command: ")
-        LOG.info(cmdlist)
-
-        _num_of_containers_before_build = self.get_number_of_created_containers()
-        command_result = run_command(cmdlist, cwd=self.working_dir)
-        self.assertEqual(command_result.process.returncode, 0)
-        _num_of_containers_after_build = self.get_number_of_created_containers()
-
-        self._verify_image_build_artifact(
-            self.built_template,
-            self.FUNCTION_LOGICAL_ID_IMAGE,
-            "ImageUri",
-            f"{self.FUNCTION_LOGICAL_ID_IMAGE.lower()}:{_tag}",
-        )
-
-        expected = {"pi": "3.14"}
-        self._verify_invoke_built_function(
-            self.built_template, self.FUNCTION_LOGICAL_ID_IMAGE, self._make_parameter_override_arg(overrides), expected
-        )
-
-        self.assertEqual(
-            _num_of_containers_before_build, _num_of_containers_after_build, "Intermediate containers are not removed"
-        )
-
-
-@skipIf(
-    # Hits public ECR pull limitation, move it to canary tests
-    (not RUN_BY_CANARY and not CI_OVERRIDE),
-    "Skip build tests on windows when running in CI unless overridden",
-)
-class TestBuildCommand_PythonFunctions_ImagesWithSharedCode(BuildIntegBase):
-    template = "template_images_with_shared_code.yaml"
-
-    EXPECTED_FILES_PROJECT_MANIFEST: Set[str] = set()
-
-    FUNCTION_LOGICAL_ID_IMAGE = "ImageFunction"
-
-    @parameterized.expand(
-        [
-            *[(runtime, "feature_phi/Dockerfile", {"phi": "1.62"}) for runtime in ["3.7", "3.8", "3.9"]],
-            *[(runtime, "feature_pi/Dockerfile", {"pi": "3.14"}) for runtime in ["3.7", "3.8", "3.9"]],
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_with_default_requirements(self, runtime, dockerfile, expected):
-        _tag = f"{random.randint(1, 100)}"
-        overrides = {
-            "Runtime": runtime,
-            "Handler": "main.handler",
-            "DockerFile": dockerfile,
-            "Tag": _tag,
-        }
-        cmdlist = self.get_command_list(use_container=False, parameter_overrides=overrides)
-
-        LOG.info("Running Command: ")
-        LOG.info(cmdlist)
-        command_result = run_command(cmdlist, cwd=self.working_dir)
-        self.assertEqual(command_result.process.returncode, 0)
-
-        self._verify_image_build_artifact(
-            self.built_template,
-            self.FUNCTION_LOGICAL_ID_IMAGE,
-            "ImageUri",
-            f"{self.FUNCTION_LOGICAL_ID_IMAGE.lower()}:{_tag}",
-        )
-
-        self._verify_invoke_built_function(
-            self.built_template, self.FUNCTION_LOGICAL_ID_IMAGE, self._make_parameter_override_arg(overrides), expected
-        )
-
-    @parameterized.expand(
-        [
-            ("feature_phi/Dockerfile", {"phi": "1.62"}),
-            ("feature_pi/Dockerfile", {"pi": "3.14"}),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_intermediate_container_deleted(self, dockerfile, expected):
-        _tag = f"{random.randint(1, 100)}"
-        overrides = {
-            "Runtime": "3.9",
-            "Handler": "main.handler",
-            "DockerFile": dockerfile,
-            "Tag": _tag,
-        }
-        cmdlist = self.get_command_list(use_container=False, parameter_overrides=overrides)
-
-        LOG.info("Running Command: ")
-        LOG.info(cmdlist)
-
-        _num_of_containers_before_build = self.get_number_of_created_containers()
-        command_result = run_command(cmdlist, cwd=self.working_dir)
-        self.assertEqual(command_result.process.returncode, 0)
-        _num_of_containers_after_build = self.get_number_of_created_containers()
-
-        self._verify_image_build_artifact(
-            self.built_template,
-            self.FUNCTION_LOGICAL_ID_IMAGE,
-            "ImageUri",
-            f"{self.FUNCTION_LOGICAL_ID_IMAGE.lower()}:{_tag}",
-        )
-
-        self._verify_invoke_built_function(
-            self.built_template, self.FUNCTION_LOGICAL_ID_IMAGE, self._make_parameter_override_arg(overrides), expected
-        )
-
-        self.assertEqual(
-            _num_of_containers_before_build, _num_of_containers_after_build, "Intermediate containers are not removed"
-        )
 
 
 @skipIf(
@@ -390,119 +188,6 @@ class TestSkipBuildingFlaggedFunctions(BuildIntegPythonBase):
             )
 
 
-@parameterized_class(
-    (
-        "template",
-        "FUNCTION_LOGICAL_ID",
-        "overrides",
-        "runtime",
-        "codeuri",
-        "use_container",
-        "check_function_only",
-        "prop",
-    ),
-    [
-        ("template.yaml", "Function", True, "python3.7", "Python", False, False, "CodeUri"),
-        ("template.yaml", "Function", True, "python3.8", "Python", False, False, "CodeUri"),
-        ("template.yaml", "Function", True, "python3.9", "Python", False, False, "CodeUri"),
-        ("template.yaml", "Function", True, "python3.10", "Python", False, False, "CodeUri"),
-        ("template.yaml", "Function", True, "python3.7", "PythonPEP600", False, False, "CodeUri"),
-        ("template.yaml", "Function", True, "python3.8", "PythonPEP600", False, False, "CodeUri"),
-        ("template.yaml", "Function", True, "python3.7", "Python", "use_container", False, "CodeUri"),
-        ("template.yaml", "Function", True, "python3.8", "Python", "use_container", False, "CodeUri"),
-        ("template.yaml", "Function", True, "python3.9", "Python", "use_container", False, "CodeUri"),
-        ("template.yaml", "Function", True, "python3.10", "Python", "use_container", False, "CodeUri"),
-    ],
-)
-class TestBuildCommand_PythonFunctions(BuildIntegPythonBase):
-    overrides = True
-    runtime = "python3.9"
-    codeuri = "Python"
-    use_container = False
-    check_function_only = False
-
-    @pytest.mark.flaky(reruns=3)
-    def test_with_default_requirements(self):
-        if self.use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-        self._test_with_default_requirements(
-            self.runtime,
-            self.codeuri,
-            self.use_container,
-            self.test_data_path,
-            do_override=self.overrides,
-            check_function_only=self.check_function_only,
-        )
-
-
-@skipIf(
-    # Hits public ECR pull limitation, move it to canary tests
-    SKIP_DOCKER_TESTS,
-    "Skip build tests that requires Docker in CI environment",
-)
-@parameterized_class(
-    (
-        "template",
-        "FUNCTION_LOGICAL_ID",
-        "overrides",
-        "runtime",
-        "codeuri",
-        "use_container",
-        "check_function_only",
-        "prop",
-    ),
-    [
-        (
-            "cdk_v1_synthesized_template_zip_image_functions.json",
-            "RandomCitiesFunction5C47A2B8",
-            False,
-            None,
-            None,
-            False,
-            True,
-            "Code",
-        ),
-    ],
-)
-class TestBuildCommand_PythonFunctions_CDK(TestBuildCommand_PythonFunctions):
-    @pytest.mark.flaky(reruns=3)
-    def test_cdk_app_with_default_requirements(self):
-        self._test_with_default_requirements(
-            self.runtime,
-            self.codeuri,
-            self.use_container,
-            self.test_data_path,
-            do_override=self.overrides,
-            check_function_only=self.check_function_only,
-        )
-
-
-class TestBuildCommand_PythonFunctions_With_Specified_Architecture(BuildIntegPythonBase):
-    template = "template_with_architecture.yaml"
-
-    @parameterized.expand(
-        [
-            ("python3.7", "Python", False, "x86_64"),
-            ("python3.8", "Python", False, "x86_64"),
-            # numpy 1.20.3 (in PythonPEP600/requirements.txt) only support python 3.7+
-            ("python3.7", "PythonPEP600", False, "x86_64"),
-            ("python3.8", "PythonPEP600", False, "x86_64"),
-            ("python3.7", "Python", "use_container", "x86_64"),
-            ("python3.8", "Python", "use_container", "x86_64"),
-            ("python3.8", "Python", False, "arm64"),
-            ("python3.8", "PythonPEP600", False, "arm64"),
-            ("python3.8", "Python", "use_container", "arm64"),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_with_default_requirements(self, runtime, codeuri, use_container, architecture):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-        self._test_with_default_requirements(
-            runtime, codeuri, use_container, self.test_data_path, architecture=architecture
-        )
-
-
 class TestBuildCommand_ErrorCases(BuildIntegBase):
     @pytest.mark.flaky(reruns=3)
     def test_unsupported_runtime(self):
@@ -515,894 +200,6 @@ class TestBuildCommand_ErrorCases(BuildIntegBase):
         self.assertEqual(1, process_execute.process.returncode)
 
         self.assertIn("Build Failed", str(process_execute.stdout))
-
-
-class TestBuildCommand_NodeFunctions(BuildIntegNodeBase):
-    @parameterized.expand(
-        [
-            ("nodejs12.x", False),
-            ("nodejs14.x", False),
-            ("nodejs16.x", False),
-            ("nodejs18.x", False),
-            ("nodejs12.x", "use_container"),
-            ("nodejs14.x", "use_container"),
-            ("nodejs16.x", "use_container"),
-            ("nodejs18.x", "use_container"),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_default_package_json(self, runtime, use_container):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-        self._test_with_default_package_json(runtime, use_container, self.test_data_path)
-
-
-class TestBuildCommand_EsbuildFunctions(BuildIntegEsbuildBase):
-    template = "template_with_metadata_esbuild.yaml"
-
-    @parameterized.expand(
-        [
-            ("nodejs14.x", "Esbuild/Node", {"main.js", "main.js.map"}, "main.lambdaHandler", False, "x86_64"),
-            ("nodejs12.x", "Esbuild/Node", {"main.js", "main.js.map"}, "main.lambdaHandler", False, "arm64"),
-            ("nodejs14.x", "Esbuild/TypeScript", {"app.js", "app.js.map"}, "app.lambdaHandler", False, "x86_64"),
-            ("nodejs12.x", "Esbuild/TypeScript", {"app.js", "app.js.map"}, "app.lambdaHandler", False, "arm64"),
-            ("nodejs14.x", "Esbuild/Node", {"main.js", "main.js.map"}, "main.lambdaHandler", "use_container", "x86_64"),
-            ("nodejs12.x", "Esbuild/Node", {"main.js", "main.js.map"}, "main.lambdaHandler", "use_container", "arm64"),
-            (
-                "nodejs14.x",
-                "Esbuild/TypeScript",
-                {"app.js", "app.js.map"},
-                "app.lambdaHandler",
-                "use_container",
-                "x86_64",
-            ),
-            (
-                "nodejs12.x",
-                "Esbuild/TypeScript",
-                {"app.js", "app.js.map"},
-                "app.lambdaHandler",
-                "use_container",
-                "arm64",
-            ),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_default_package_json(
-        self, runtime, code_uri, expected_files, handler, use_container, architecture
-    ):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-        self._test_with_default_package_json(runtime, use_container, code_uri, expected_files, handler, architecture)
-
-
-@skipIf(
-    ((IS_WINDOWS and RUNNING_ON_CI) and not CI_OVERRIDE),
-    "Skip build tests on windows when running in CI unless overridden",
-)
-@parameterized_class(
-    ("template",),
-    [
-        ("esbuild_templates/template_with_metadata_node_options.yaml",),
-        ("esbuild_templates/template_with_metadata_global_node_options.yaml",),
-    ],
-)
-class TestBuildCommand_EsbuildFunctionProperties(BuildIntegEsbuildBase):
-    @parameterized.expand(
-        [
-            ("nodejs16.x", "../Esbuild/TypeScript", "app.lambdaHandler", "x86_64"),
-            ("nodejs18.x", "../Esbuild/TypeScript", "app.lambdaHandler", "x86_64"),
-            ("nodejs16.x", "../Esbuild/TypeScript", "nested/function/app.lambdaHandler", "x86_64"),
-            ("nodejs18.x", "../Esbuild/TypeScript", "nested/function/app.lambdaHandler", "x86_64"),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_environment_generates_sourcemap(self, runtime, code_uri, handler, architecture):
-        overrides = {
-            "runtime": runtime,
-            "code_uri": code_uri,
-            "handler": handler,
-            "architecture": architecture,
-        }
-        self._test_with_various_properties(overrides)
-
-
-class TestBuildCommand_NodeFunctions_With_Specified_Architecture(BuildIntegNodeBase):
-    template = "template_with_architecture.yaml"
-
-    @parameterized.expand(
-        [
-            ("nodejs12.x", False, "x86_64"),
-            ("nodejs14.x", False, "x86_64"),
-            ("nodejs16.x", False, "x86_64"),
-            ("nodejs18.x", False, "x86_64"),
-            ("nodejs12.x", "use_container", "x86_64"),
-            ("nodejs14.x", "use_container", "x86_64"),
-            ("nodejs16.x", "use_container", "x86_64"),
-            ("nodejs18.x", "use_container", "x86_64"),
-            ("nodejs12.x", False, "arm64"),
-            ("nodejs14.x", False, "arm64"),
-            ("nodejs16.x", False, "arm64"),
-            ("nodejs18.x", False, "arm64"),
-            ("nodejs12.x", "use_container", "arm64"),
-            ("nodejs14.x", "use_container", "arm64"),
-            ("nodejs16.x", "use_container", "arm64"),
-            ("nodejs18.x", "use_container", "arm64"),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_default_package_json(self, runtime, use_container, architecture):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-        self._test_with_default_package_json(runtime, use_container, self.test_data_path, architecture)
-
-
-class TestBuildCommand_RubyFunctions(BuildIntegRubyBase):
-    @parameterized.expand(["ruby2.7"])
-    @pytest.mark.flaky(reruns=3)
-    @skipIf(SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD, SKIP_DOCKER_MESSAGE)
-    def test_building_ruby_in_container(self, runtime):
-        self._test_with_default_gemfile(runtime, "use_container", "Ruby", self.test_data_path)
-
-    @parameterized.expand(["ruby2.7"])
-    @pytest.mark.flaky(reruns=3)
-    def test_building_ruby_in_process(self, runtime):
-        self._test_with_default_gemfile(runtime, False, "Ruby", self.test_data_path)
-
-
-class TestBuildCommand_RubyFunctions_With_Architecture(BuildIntegRubyBase):
-    template = "template_with_architecture.yaml"
-
-    @parameterized.expand(["ruby2.7", ("ruby2.7", "arm64")])
-    @pytest.mark.flaky(reruns=3)
-    @skipIf(SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD, SKIP_DOCKER_MESSAGE)
-    def test_building_ruby_in_container_with_specified_architecture(self, runtime, architecture="x86_64"):
-        self._test_with_default_gemfile(runtime, "use_container", "Ruby", self.test_data_path, architecture)
-
-    @parameterized.expand(["ruby2.7", ("ruby2.7", "arm64")])
-    @pytest.mark.flaky(reruns=3)
-    def test_building_ruby_in_process_with_specified_architecture(self, runtime, architecture="x86_64"):
-        self._test_with_default_gemfile(runtime, False, "Ruby", self.test_data_path, architecture)
-
-
-class TestBuildCommand_RubyFunctionsWithGemfileInTheRoot(BuildIntegRubyBase):
-    """
-    Tests use case where Gemfile will present in the root of the project folder.
-    This doesn't apply to containerized build, since it copies only the function folder to the container
-    """
-
-    @parameterized.expand([("ruby2.7")])
-    @pytest.mark.flaky(reruns=3)
-    def test_building_ruby_in_process_with_root_gemfile(self, runtime):
-        self._prepare_application_environment()
-        self._test_with_default_gemfile(runtime, False, "RubyWithRootGemfile", self.working_dir)
-
-    def _prepare_application_environment(self):
-        """
-        Create an application environment where Gemfile will be in the root folder of the app;
-        ├── RubyWithRootGemfile
-        │   └── app.rb
-        ├── Gemfile
-        └── template.yaml
-        """
-        # copy gemfile to the root of the project
-        shutil.copyfile(Path(self.template_path).parent.joinpath("Gemfile"), Path(self.working_dir).joinpath("Gemfile"))
-        # copy function source code in its folder
-        osutils.copytree(
-            Path(self.template_path).parent.joinpath("RubyWithRootGemfile"),
-            Path(self.working_dir).joinpath("RubyWithRootGemfile"),
-        )
-        # copy template to the root folder
-        shutil.copyfile(Path(self.template_path), Path(self.working_dir).joinpath("template.yaml"))
-        # update template path with new location
-        self.template_path = str(Path(self.working_dir).joinpath("template.yaml"))
-
-
-class TestBuildCommand_Java(BuildIntegJavaBase):
-    EXPECTED_FILES_PROJECT_MANIFEST_GRADLE = {"aws", "lib", "META-INF"}
-    EXPECTED_FILES_PROJECT_MANIFEST_MAVEN = {"aws", "lib"}
-    EXPECTED_GRADLE_DEPENDENCIES = {"annotations-2.1.0.jar", "aws-lambda-java-core-1.1.0.jar"}
-    EXPECTED_MAVEN_DEPENDENCIES = {
-        "software.amazon.awssdk.annotations-2.1.0.jar",
-        "com.amazonaws.aws-lambda-java-core-1.1.0.jar",
-    }
-
-    FUNCTION_LOGICAL_ID = "Function"
-    USING_GRADLE_PATH = os.path.join("Java", "gradle")
-    USING_GRADLEW_PATH = os.path.join("Java", "gradlew")
-    USING_GRADLE_KOTLIN_PATH = os.path.join("Java", "gradle-kotlin")
-    USING_MAVEN_PATH = os.path.join("Java", "maven")
-
-    @parameterized.expand(
-        [
-            ("java8", "8", USING_GRADLE_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java8", "8", USING_GRADLEW_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            (
-                "java8",
-                "8",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            ("java8", "8", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES),
-            ("java8.al2", "8", USING_GRADLE_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            (
-                "java8.al2",
-                "8",
-                USING_GRADLEW_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            (
-                "java8.al2",
-                "8",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            ("java8.al2", "8", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES),
-            ("java11", "11", USING_GRADLE_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java11", "11", USING_GRADLEW_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            (
-                "java11",
-                "11",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            ("java11", "11", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES),
-            ("java17", "17", USING_GRADLE_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java17", "17", USING_GRADLEW_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            (
-                "java17",
-                "17",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            ("java17", "17", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES),
-        ]
-    )
-    @skipIf(SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD, SKIP_DOCKER_MESSAGE)
-    @pytest.mark.flaky(reruns=3)
-    def test_building_java_in_container(
-        self, runtime, runtime_version, code_path, expected_files, expected_dependencies
-    ):
-        self._test_with_building_java(
-            runtime,
-            os.path.join(code_path, runtime_version),
-            expected_files,
-            expected_dependencies,
-            "use_container",
-            self.test_data_path,
-        )
-
-    @parameterized.expand(
-        [
-            ("java8", USING_GRADLE_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java8", USING_GRADLEW_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java8", USING_GRADLE_KOTLIN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java8", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES),
-            ("java8.al2", USING_GRADLE_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java8.al2", USING_GRADLEW_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            (
-                "java8.al2",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-            ),
-            ("java8.al2", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_java8_in_process(self, runtime, code_path, expected_files, expected_dependencies):
-        self._test_with_building_java(
-            runtime, os.path.join(code_path, "8"), expected_files, expected_dependencies, False, self.test_data_path
-        )
-
-    @parameterized.expand(
-        [
-            ("java11", USING_GRADLE_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java11", USING_GRADLEW_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java11", USING_GRADLE_KOTLIN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java11", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_java11_in_process(self, runtime, code_path, expected_files, expected_dependencies):
-        self._test_with_building_java(
-            runtime, os.path.join(code_path, "11"), expected_files, expected_dependencies, False, self.test_data_path
-        )
-
-    @parameterized.expand(
-        [
-            ("java17", USING_GRADLE_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java17", USING_GRADLEW_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java17", USING_GRADLE_KOTLIN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_GRADLE, EXPECTED_GRADLE_DEPENDENCIES),
-            ("java17", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_java17_in_process(self, runtime, code_path, expected_files, expected_dependencies):
-        self._test_with_building_java(
-            runtime, os.path.join(code_path, "17"), expected_files, expected_dependencies, False, self.test_data_path
-        )
-
-
-@skipIf(
-    (IS_WINDOWS and not CI_OVERRIDE),
-    "Skip build tests on windows when running in CI unless overridden",
-)
-class TestBuildCommand_Java_With_Specified_Architecture(BuildIntegJavaBase):
-    template = "template_with_architecture.yaml"
-    EXPECTED_FILES_PROJECT_MANIFEST_GRADLE = {"aws", "lib", "META-INF"}
-    EXPECTED_FILES_PROJECT_MANIFEST_MAVEN = {"aws", "lib"}
-    EXPECTED_GRADLE_DEPENDENCIES = {"annotations-2.1.0.jar", "aws-lambda-java-core-1.1.0.jar"}
-    EXPECTED_MAVEN_DEPENDENCIES = {
-        "software.amazon.awssdk.annotations-2.1.0.jar",
-        "com.amazonaws.aws-lambda-java-core-1.1.0.jar",
-    }
-
-    FUNCTION_LOGICAL_ID = "Function"
-    USING_GRADLE_PATH = os.path.join("Java", "gradle")
-    USING_GRADLEW_PATH = os.path.join("Java", "gradlew")
-    USING_GRADLE_KOTLIN_PATH = os.path.join("Java", "gradle-kotlin")
-    USING_MAVEN_PATH = os.path.join("Java", "maven")
-
-    @parameterized.expand(
-        [
-            (
-                "java8.al2",
-                "8",
-                USING_GRADLE_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java8.al2",
-                "8",
-                USING_GRADLEW_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java8.al2",
-                "8",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java8.al2",
-                "8",
-                USING_MAVEN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                EXPECTED_MAVEN_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java11",
-                "11",
-                USING_GRADLE_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java11",
-                "11",
-                USING_GRADLEW_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java11",
-                "11",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java11",
-                "11",
-                USING_MAVEN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                EXPECTED_MAVEN_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java17",
-                "17",
-                USING_GRADLE_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java17",
-                "17",
-                USING_GRADLEW_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java17",
-                "17",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java17",
-                "17",
-                USING_MAVEN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                EXPECTED_MAVEN_DEPENDENCIES,
-                "arm64",
-            ),
-        ]
-    )
-    @skipIf(SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD, SKIP_DOCKER_MESSAGE)
-    @pytest.mark.flaky(reruns=3)
-    def test_building_java_in_container_with_arm64_architecture(
-        self, runtime, runtime_version, code_path, expected_files, expected_dependencies, architecture
-    ):
-        self._test_with_building_java(
-            runtime,
-            os.path.join(code_path, runtime_version),
-            expected_files,
-            expected_dependencies,
-            "use_container",
-            self.test_data_path,
-            architecture,
-        )
-
-    @parameterized.expand(
-        [
-            (
-                "java8.al2",
-                USING_GRADLE_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java8.al2",
-                USING_GRADLEW_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java8.al2",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java8.al2",
-                USING_MAVEN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_MAVEN,
-                EXPECTED_MAVEN_DEPENDENCIES,
-                "arm64",
-            ),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_java8_in_process_with_arm_architecture(
-        self, runtime, code_path, expected_files, expected_dependencies, architecture
-    ):
-        self._test_with_building_java(
-            runtime,
-            os.path.join(code_path, "8"),
-            expected_files,
-            expected_dependencies,
-            False,
-            self.test_data_path,
-            architecture,
-        )
-
-    @parameterized.expand(
-        [
-            (
-                "java11",
-                USING_GRADLE_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java11",
-                USING_GRADLEW_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java11",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            ("java11", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES, "arm64"),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_java11_in_process_with_arm_architecture(
-        self, runtime, code_path, expected_files, expected_dependencies, architecture
-    ):
-        self._test_with_building_java(
-            runtime,
-            os.path.join(code_path, "11"),
-            expected_files,
-            expected_dependencies,
-            False,
-            self.test_data_path,
-            architecture,
-        )
-
-    @parameterized.expand(
-        [
-            (
-                "java17",
-                USING_GRADLE_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java17",
-                USING_GRADLEW_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            (
-                "java17",
-                USING_GRADLE_KOTLIN_PATH,
-                EXPECTED_FILES_PROJECT_MANIFEST_GRADLE,
-                EXPECTED_GRADLE_DEPENDENCIES,
-                "arm64",
-            ),
-            ("java17", USING_MAVEN_PATH, EXPECTED_FILES_PROJECT_MANIFEST_MAVEN, EXPECTED_MAVEN_DEPENDENCIES, "arm64"),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_java17_in_process_with_arm_architecture(
-        self, runtime, code_path, expected_files, expected_dependencies, architecture
-    ):
-        self._test_with_building_java(
-            runtime,
-            os.path.join(code_path, "17"),
-            expected_files,
-            expected_dependencies,
-            False,
-            self.test_data_path,
-            architecture,
-        )
-
-
-class TestBuildCommand_Dotnet_cli_package(BuildIntegBase):
-    FUNCTION_LOGICAL_ID = "Function"
-    EXPECTED_FILES_PROJECT_MANIFEST = {
-        "Amazon.Lambda.APIGatewayEvents.dll",
-        "Amazon.Lambda.Core.dll",
-        "HelloWorld.runtimeconfig.json",
-        "Amazon.Lambda.Serialization.Json.dll",
-        "Newtonsoft.Json.dll",
-        "HelloWorld.deps.json",
-        "HelloWorld.dll",
-    }
-
-    EXPECTED_FILES_PROJECT_MANIFEST_PROVIDED = {
-        "bootstrap",
-    }
-
-    @parameterized.expand(
-        [
-            ("dotnetcore3.1", "Dotnetcore3.1", None),
-            ("dotnet6", "Dotnet6", None),
-            ("dotnetcore3.1", "Dotnetcore3.1", "debug"),
-            ("dotnet6", "Dotnet6", "debug"),
-            ("provided.al2", "Dotnet7", None),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_dotnetcore_in_process(self, runtime, code_uri, mode, architecture="x86_64"):
-        # dotnet7 requires docker to build the function
-        if code_uri == "Dotnet7" and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-        overrides = {
-            "Runtime": runtime,
-            "CodeUri": code_uri,
-            "Handler": "HelloWorld::HelloWorld.Function::FunctionHandler",
-            "Architectures": architecture,
-        }
-
-        if runtime == "provided.al2":
-            self.template_path = self.template_path.replace("template.yaml", "template_build_method_dotnet_7.yaml")
-
-        cmdlist = self.get_command_list(use_container=False, parameter_overrides=overrides)
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        LOG.info("Running with SAM_BUILD_MODE={}".format(mode))
-
-        newenv = os.environ.copy()
-        if mode:
-            newenv["SAM_BUILD_MODE"] = mode
-
-        command_result = run_command(cmdlist, cwd=self.working_dir, env=newenv)
-        self.assertEqual(command_result.process.returncode, 0)
-
-        self._verify_built_artifact(
-            self.default_build_dir,
-            self.FUNCTION_LOGICAL_ID,
-            self.EXPECTED_FILES_PROJECT_MANIFEST
-            if runtime != "provided.al2"
-            else self.EXPECTED_FILES_PROJECT_MANIFEST_PROVIDED,
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "OtherRelativePathResource",
-            "BodyS3Location",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(self.test_data_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "GlueResource",
-            "Command.ScriptLocation",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(self.test_data_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        expected = "{'message': 'Hello World'}"
-        if not SKIP_DOCKER_TESTS:
-            self._verify_invoke_built_function(
-                self.built_template, self.FUNCTION_LOGICAL_ID, self._make_parameter_override_arg(overrides), expected
-            )
-            self.verify_docker_container_cleanedup(runtime)
-
-    @parameterized.expand(
-        [
-            ("dotnetcore3.1", "Dotnetcore3.1", None),
-            ("dotnet6", "Dotnet6", None),
-            ("dotnetcore3.1", "Dotnetcore3.1", "debug"),
-            ("dotnet6", "Dotnet6", "debug"),
-            # force to run tests on arm64 machines may cause dotnet7 test failing
-            # because Native AOT Lambda functions require the host and lambda architectures to match
-            ("provided.al2", "Dotnet7", None),
-        ]
-    )
-    @skipIf(SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD, SKIP_DOCKER_MESSAGE)
-    @pytest.mark.flaky(reruns=3)
-    def test_dotnetcore_in_container_mount_with_write_explicit(self, runtime, code_uri, mode, architecture="x86_64"):
-        overrides = {
-            "Runtime": runtime,
-            "CodeUri": code_uri,
-            "Handler": "HelloWorld::HelloWorld.Function::FunctionHandler",
-            "Architectures": architecture,
-        }
-
-        if runtime == "provided.al2":
-            self.template_path = self.template_path.replace("template.yaml", "template_build_method_dotnet_7.yaml")
-
-        # test with explicit mount_with_write flag
-        cmdlist = self.get_command_list(use_container=True, parameter_overrides=overrides, mount_with=MountMode.WRITE)
-        # env vars needed for testing unless set by dotnet images on public.ecr.aws
-        cmdlist += ["--container-env-var", "DOTNET_CLI_HOME=/tmp/dotnet"]
-        cmdlist += ["--container-env-var", "XDG_DATA_HOME=/tmp/xdg"]
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        LOG.info("Running with SAM_BUILD_MODE={}".format(mode))
-
-        newenv = os.environ.copy()
-        if mode:
-            newenv["SAM_BUILD_MODE"] = mode
-
-        command_result = run_command(cmdlist, cwd=self.working_dir, env=newenv)
-        self.assertEqual(command_result.process.returncode, 0)
-
-        self._verify_built_artifact(
-            self.default_build_dir,
-            self.FUNCTION_LOGICAL_ID,
-            self.EXPECTED_FILES_PROJECT_MANIFEST
-            if runtime != "provided.al2"
-            else self.EXPECTED_FILES_PROJECT_MANIFEST_PROVIDED,
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "OtherRelativePathResource",
-            "BodyS3Location",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(self.test_data_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "GlueResource",
-            "Command.ScriptLocation",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(self.test_data_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        expected = "{'message': 'Hello World'}"
-        self._verify_invoke_built_function(
-            self.built_template, self.FUNCTION_LOGICAL_ID, self._make_parameter_override_arg(overrides), expected
-        )
-        self.verify_docker_container_cleanedup(runtime)
-
-    @parameterized.expand(
-        [
-            ("dotnetcore3.1", "Dotnetcore3.1", None),
-            ("dotnet6", "Dotnet6", None),
-            ("dotnetcore3.1", "Dotnetcore3.1", "debug"),
-            ("dotnet6", "Dotnet6", "debug"),
-            # force to run tests on arm64 machines may cause dotnet7 test failing
-            # because Native AOT Lambda functions require the host and lambda architectures to match
-            ("provided.al2", "Dotnet7", None),
-        ]
-    )
-    @skipIf(SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD, SKIP_DOCKER_MESSAGE)
-    @pytest.mark.flaky(reruns=3)
-    def test_dotnetcore_in_container_mount_with_write_interactive(
-        self,
-        runtime,
-        code_uri,
-        mode,
-        architecture="x86_64",
-    ):
-        overrides = {
-            "Runtime": runtime,
-            "CodeUri": code_uri,
-            "Handler": "HelloWorld::HelloWorld.Function::FunctionHandler",
-            "Architectures": architecture,
-        }
-
-        if runtime == "provided.al2":
-            self.template_path = self.template_path.replace("template.yaml", "template_build_method_dotnet_7.yaml")
-
-        # test without explicit mount_with_write flag
-        cmdlist = self.get_command_list(use_container=True, parameter_overrides=overrides)
-        # env vars needed for testing unless set by dotnet images on public.ecr.aws
-        cmdlist += ["--container-env-var", "DOTNET_CLI_HOME=/tmp/dotnet"]
-        cmdlist += ["--container-env-var", "XDG_DATA_HOME=/tmp/xdg"]
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        LOG.info("Running with SAM_BUILD_MODE={}".format(mode))
-
-        # mock user input to mount with write
-        user_click_confirm_input = "y"
-        command_result = run_command_with_input(cmdlist, user_click_confirm_input.encode(), cwd=self.working_dir)
-        self.assertEqual(command_result.process.returncode, 0)
-
-        self._verify_built_artifact(
-            self.default_build_dir,
-            self.FUNCTION_LOGICAL_ID,
-            self.EXPECTED_FILES_PROJECT_MANIFEST
-            if runtime != "provided.al2"
-            else self.EXPECTED_FILES_PROJECT_MANIFEST_PROVIDED,
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "OtherRelativePathResource",
-            "BodyS3Location",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(self.test_data_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "GlueResource",
-            "Command.ScriptLocation",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(self.test_data_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        expected = "{'message': 'Hello World'}"
-        self._verify_invoke_built_function(
-            self.built_template, self.FUNCTION_LOGICAL_ID, self._make_parameter_override_arg(overrides), expected
-        )
-        self.verify_docker_container_cleanedup(runtime)
-
-    @parameterized.expand([("dotnetcore3.1", "Dotnetcore3.1"), ("dotnet6", "Dotnet6")])
-    @skipIf(SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD, SKIP_DOCKER_MESSAGE)
-    @pytest.mark.flaky(reruns=3)
-    def test_must_fail_on_container_mount_without_write_interactive(self, runtime, code_uri):
-        use_container = True
-        overrides = {
-            "Runtime": runtime,
-            "CodeUri": code_uri,
-            "Handler": "HelloWorld::HelloWorld.Function::FunctionHandler",
-        }
-        cmdlist = self.get_command_list(use_container=use_container, parameter_overrides=overrides)
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        # mock user input to not allow mounting with write
-        user_click_confirm_input = "N"
-        process_execute = run_command_with_input(cmdlist, user_click_confirm_input.encode())
-
-        # Must error out, because mounting with write is not allowed
-        self.assertEqual(process_execute.process.returncode, 1)
-
-    def _verify_built_artifact(self, build_dir, function_logical_id, expected_files):
-        self.assertTrue(build_dir.exists(), "Build directory should be created")
-
-        build_dir_files = os.listdir(str(build_dir))
-        self.assertIn("template.yaml", build_dir_files)
-        self.assertIn(function_logical_id, build_dir_files)
-
-        template_path = build_dir.joinpath("template.yaml")
-        resource_artifact_dir = build_dir.joinpath(function_logical_id)
-
-        # Make sure the template has correct CodeUri for resource
-        self._verify_resource_property(str(template_path), function_logical_id, "CodeUri", function_logical_id)
-
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        actual_files = all_artifacts.intersection(expected_files)
-        self.assertEqual(actual_files, expected_files)
-
-
-class TestBuildCommand_Go_Modules(BuildIntegGoBase):
-    @parameterized.expand([("go1.x", "Go", None, False), ("go1.x", "Go", "debug", True)])
-    @pytest.mark.flaky(reruns=3)
-    def test_building_go(self, runtime, code_uri, mode, use_container):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-
-        self._test_with_go(runtime, code_uri, mode, self.test_data_path, use_container=use_container)
-
-
-class TestBuildCommand_Go_Modules_With_Specified_Architecture(BuildIntegGoBase):
-    template = "template_with_architecture.yaml"
-
-    @parameterized.expand(
-        [
-            ("go1.x", "Go", None, "x86_64"),
-            ("go1.x", "Go", "debug", "x86_64"),
-            ("go1.x", "Go", None, "arm64"),
-            ("go1.x", "Go", "debug", "arm64"),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_go(self, runtime, code_uri, mode, architecture):
-        self._test_with_go(runtime, code_uri, mode, self.test_data_path, architecture)
-
-    @parameterized.expand([("go1.x", "Go", "unknown_architecture")])
-    @pytest.mark.flaky(reruns=3)
-    def test_go_must_fail_with_unknown_architecture(self, runtime, code_uri, architecture):
-        overrides = {"Runtime": runtime, "CodeUri": code_uri, "Handler": "hello-world", "Architectures": architecture}
-        cmdlist = self.get_command_list(parameter_overrides=overrides)
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        process_execute = run_command(cmdlist, cwd=self.working_dir)
-
-        # Must error out, because container builds are not supported
-        self.assertEqual(process_execute.process.returncode, 1)
 
 
 class TestBuildCommand_SingleFunctionBuilds(BuildIntegBase):
@@ -1702,88 +499,6 @@ class TestBuildCommand_LayerBuilds(BuildIntegBase):
 
     def _get_python_version(self):
         return "python{}.{}".format(sys.version_info.major, sys.version_info.minor)
-
-
-@parameterized_class(
-    ("template", "is_nested_parent"),
-    [
-        (os.path.join("nested-parent", "template-parent.yaml"), "is_nested_parent"),
-        ("template.yaml", False),
-    ],
-)
-class TestBuildCommand_ProvidedFunctions(BuildIntegProvidedBase):
-    # Test Suite for runtime: provided and where selection of the build workflow is implicitly makefile builder
-    # if the makefile is present.
-    @parameterized.expand(
-        [
-            ("provided", False, None),
-            ("provided", "use_container", "Makefile-container"),
-            ("provided.al2", False, None),
-            ("provided.al2", "use_container", "Makefile-container"),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_Makefile(self, runtime, use_container, manifest):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-        self._test_with_Makefile(runtime, use_container, manifest)
-
-
-@parameterized_class(
-    ("template", "is_nested_parent"),
-    [
-        (os.path.join("nested-parent", "template-parent.yaml"), "is_nested_parent"),
-        ("template.yaml", False),
-    ],
-)
-class TestBuildCommand_ProvidedFunctions_With_Specified_Architecture(BuildIntegProvidedBase):
-    # Test Suite for runtime: provided and where selection of the build workflow is implicitly makefile builder
-    # if the makefile is present.
-    @parameterized.expand(
-        [
-            ("provided", False, None, "x86_64"),
-            ("provided", "use_container", "Makefile-container", "x86_64"),
-            ("provided.al2", False, None, "x86_64"),
-            ("provided.al2", "use_container", "Makefile-container", "x86_64"),
-            ("provided", False, None, "arm64"),
-            ("provided", "use_container", "Makefile-container", "arm64"),
-            ("provided.al2", False, None, "arm64"),
-            ("provided.al2", "use_container", "Makefile-container", "arm64"),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_Makefile(self, runtime, use_container, manifest, architecture):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-        self._test_with_Makefile(runtime, use_container, manifest, architecture)
-
-
-@parameterized_class(
-    ("template", "code_uri", "is_nested_parent"),
-    [
-        ("custom_build_with_custom_root_project_path.yaml", "empty_src_code", False),
-        ("custom_build_with_custom_make_file_path.yaml", "provided_src_code_without_makefile", False),
-        ("custom_build_with_custom_working_dir.yaml", "custom_working_dir_src_code", False),
-        ("custom_build_with_custom_root_project_path_and_custom_makefile_path.yaml", "empty_src_code", False),
-        (
-            "custom_build_with_custom_root_project_path_custom_makefile_path_and_custom_working_dir.yaml",
-            "empty_src_code",
-            False,
-        ),
-    ],
-)
-class TestBuildCommand_ProvidedFunctionsWithCustomMetadata(BuildIntegProvidedBase):
-    # Test Suite for runtime: provided and where selection of the build workflow is implicitly makefile builder
-    # if the makefile is present.
-    @parameterized.expand(
-        [
-            ("provided", False, None),
-            ("provided.al2", False, None),
-        ]
-    )
-    @pytest.mark.flaky(reruns=3)
-    def test_building_Makefile(self, runtime, use_container, manifest):
-        self._test_with_Makefile(runtime, use_container, manifest)
 
 
 @skipIf(
